Refactor API route resolvers and consolidate route directory logic
